refatorando o codigo quase que todo. excluindo a pasta controllers e reafazendo o routes de todos os modulos, reafazendo tambem quase todos os service.

+++ backend/src/modules/category/service.ts
@@ -1,28 +1,38 @@
 import { db } from '../../db';
 import { tablecategories } from '../../db/schema/eccomerce/categories';
-import { eq } from 'drizzle-orm';
+import { eq, and } from 'drizzle-orm';
 
 interface CreateCategoryInput {
     name: string
     description: string
 }
 
-export const createCategory = async (category: CreateCategoryInput) => {
+export const createCategory = async (userId: string, category: CreateCategoryInput) => {
 
-    const verifyName = await db.select().from(tablecategories).where(eq(tablecategories.name, category.name))
+    const verifyName = await db
+        .select()
+        .from(tablecategories)
+        .where(eq(tablecategories.name, category.name))
+
     if (verifyName.length > 0) {
-        throw new Error ("Category ja existe")
+        throw new Error("Category já existe")
     }
 
-    const create = await db.insert(tablecategories).values(category).returning()
+    const create = await db.insert(tablecategories).values({
+        name: category.name,
+        description: category.description,
+        userId: userId
+    }).returning()
 
     if (!create || create.length === 0) {
-        throw new Error ("Category não foi criada")
+        throw new Error("Category não foi criada")
     }
-    return create[0] ?? null
+
+    return create[0]
 }
 
-export const updateCategory = async (id: number, category: Partial<CreateCategoryInput>) => {
+
+export const updateCategory = async (id: string, category: Partial<CreateCategoryInput>) => {
     const existing = await db.select().from(tablecategories).where(eq(tablecategories.id,id))
     if (!existing || existing.length === 0) {
         throw new Error("Category nao existe")
@@ -34,8 +44,8 @@ export const updateCategory = async (id: number, category: Partial<CreateCategor
     return update[0] ?? null
 }
 
-export const getCategoryById = async (id: number) => {
-    const result = await db.select().from(tablecategories).where(eq(tablecategories.id,id))
+export const getCategoryById = async (userId:string,id: string) => {
+    const result = await db.select().from(tablecategories).where(and(eq(tablecategories.id,id),eq(tablecategories.userId,userId)))
     if (!result || result.length === 0) {
         throw new Error("Category não foi encontrada")
     }
@@ -55,11 +65,11 @@ export const getCategoryByName = async (name:string) => {
         throw new Error("Category não foi encontrada")
     }
     return res[0] ?? null
-    
 }
 
-export const deletCategory = async (id: number) => {
-    const delet = await db.delete(tablecategories).where(eq(tablecategories.id,id)).returning()
+
+export const deletCategory = async (userId: string, id: string) => {
+    const delet = await db.delete(tablecategories).where(and(eq(tablecategories.id,id),eq(tablecategories.userId,userId))).returning()
     if (!delet || delet.length === 0) {
         throw new Error("Category não foi deletada")
     }
